requirements.txt
python-dotenv==0.9.1
python-telegram-bot==13.11
requests==2.20.1


